requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
# paddleocr  # Optional: Better OCR (requires: pip install paddlepaddle-gpu OR paddlepaddle)
# layoutparser==0.3.4  # Optional: Advanced layout parsing
# torch==2.1.0  # Only needed for layoutparser/paddleocr
# torchvision==0.16.0  # Only needed for layoutparser

# Audio Processing
openai-whisper==20231117  # Open-source ASR
pyannote.audio==3.1.1
noisereduce==3.0.0

# NLP & Embeddings (100% Local - NO API CALLS)
sentence-transformers==2.2.2
transformers==4.36.0
spacy==3.7.2
yake==0.4.8
# REMOVED: openai==1.6.1 (paid API)
# REMOVED: anthropic (paid API)

# Local LLM Inference (Kaggle-compatible models)
llama-cpp-python==0.2.27  # For GGUF quantized models (CPU/GPU)
huggingface-hub==0.19.4  # For downloading models

# Vector Store & Search
#faiss-cpu==1.7.4
faiss-gpu==1.7.2  # Uncomment for GPU support
# REMOVED: elasticsearch (not needed for MVP)
# REMOVED: pymilvus (not needed for MVP)

# Retrieval & Reranking (100% Local)
rank-bm25==0.2.2

# Training & Finetuning
peft==0.7.1  # Parameter-Efficient Fine-Tuning (LoRA)
bitsandbytes==0.41.3  # 4-bit quantization for 4GB GPU
accelerate==0.25.0  # Distributed training utilities
datasets==2.15.0  # HuggingFace datasets
trl==0.7.4  # Transformer Reinforcement Learning (for SFT)

# Hyperparameter Optimization
optuna==3.5.0  # Bayesian optimization

# Evaluation Metrics
rouge-score==0.1.2  # ROUGE metrics for summaries
bert-score==0.3.13  # BERTScore for semantic similarity

# Web Search (Free, No API Key)
duckduckgo-search==4.1.1  # Free web search

# Utilities
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
requests==2.31.0

# Export
genanki==0.13.1

# MCP Server
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.7.1

